requirements.txt
numpy
opencv-python
keras-nightly
pillow
streamlit>=1.30